requirements.txt
alembic==0.7.5.post2
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
blinker==1.3
celery==3.1.18
cma==1.1.6
eventlet==0.17.4
Flask==0.10.1
Flask-Bootstrap==3.3.2.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-SSLify==0.1.5
Flask-WTF==0.11
greenlet==0.4.7
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.26
lmfit==0.8.3
Mako==1.0.1
MarkupSafe==0.23
numpy==1.8.2
psycopg2==2.6
python-magic==0.4.6
pytz==2015.2
scipy==0.15.1
SQLAlchemy==1.0.2
Werkzeug==0.10.4
WTForms==2.0.2
